Fix warning: implicit conversion changes signedness
visp/tutorial/image/tutorial-canny-hsv.cpp: In function ‘int main(int, const char**)’:
visp/tutorial/image/tutorial-canny-hsv.cpp:262:31: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  262 |   int height = Iload.getRows()/4;
      |                ~~~~~~~~~~~~~~~^~
visp/tutorial/image/tutorial-canny-hsv.cpp:263:30: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  263 |   int width = Iload.getCols()/4;
      |               ~~~~~~~~~~~~~~~^~
visp/tutorial/image/tutorial-canny-hsv.cpp:264:34: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  264 |   int midHeight = Iload.getRows()/2;
      |                   ~~~~~~~~~~~~~~~^~
visp/tutorial/image/tutorial-canny-hsv.cpp:265:33: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  265 |   int midWidth = Iload.getCols()/2;
      |                  ~~~~~~~~~~~~~~~^~

+++ tutorial/image/tutorial-canny-hsv.cpp
@@ -259,10 +259,10 @@ int main(int argc, const char *argv[])
   }
 
   vpImage<bool> mask(Iload.getRows(), Iload.getCols(), false);
-  int height = Iload.getRows()/4;
-  int width = Iload.getCols()/4;
-  int midHeight = Iload.getRows()/2;
-  int midWidth = Iload.getCols()/2;
+  int height = static_cast<int>(Iload.getRows())/4;
+  int width = static_cast<int>(Iload.getCols())/4;
+  int midHeight = static_cast<int>(Iload.getRows())/2;
+  int midWidth = static_cast<int>(Iload.getCols())/2;
   for (int r = 0; r < height; ++r) {
     for (int c = 0; c < width; ++c) {
       mask[midHeight - r][midWidth - c] = true;
